requirements.txt
aniso8601==3.0.0
astroid==1.6.3
attrs==17.4.0
Django==2.0.4
django-environ==0.4.4
django-filter==1.1.0
django-filters==0.2.1
django-graphql-jwt==0.1.9
graphene==2.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.1.0
pluggy==0.6.0
promise==2.1
py==1.5.3
PyJWT==1.6.1
pylint==1.8.4
pylint-django==0.9.4
pylint-plugin-utils==0.2.6
pytest==3.5.0
pytz==2018.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
wrapt==1.10.11
